pyproject.toml
[project]
name = "langchain-simple-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.18",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.28",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "wikipedia>=1.4.0",
]
